tqma95xxsa: configure WDOG_ANY as Open Drain
This enables a working reset in the way implemented on NXP EVK.
Relying on POR_B for an reset button implementation does not
work and forces the system in an endless boot loop.

Since there are PU resistors in the relevant connections on the SoM
we need to drive the pin as Open Drain. Therefore the pin config
in pad configuration register has to be changed from its default
to sane settings for Open Drain:

- only PU, no PD
- slew rate to fast (analogue to I2C pins)
- no drive strength needed
- no hysteresis

To simplify application of the settings in various use cases
the pin configuration is factored out to a function.

Fixes: 3b785bac50ff ("tqma95xxsa: add initial version")

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>